+++ src/euphorie/content/fti.py
@@ -1,6 +1,9 @@
+from Acquisition import aq_base
 from zope.interface import Interface
 from zope.component import queryMultiAdapter
+from zope.component import queryUtility
 from plone.dexterity.fti import DexterityFTI
+from plone.dexterity.interfaces import IDexterityFTI
 
 
 class IConstructionFilter(Interface):
@@ -40,3 +43,14 @@ class ConditionalDexterityFTI(DexterityFTI):
             return filter.allowed()
 
         return True
+
+
+def check_fti_paste_allowed(container, obj):
+    portal_type = getattr(aq_base(obj), 'portal_type', None)
+    if not portal_type:
+        raise ValueError('Can only paste portal content.')
+    fti = queryUtility(IDexterityFTI, name=portal_type)
+    if fti is None:
+        raise ValueError('Can not paste non-dexterity content.')
+    if not fti.isConstructionAllowed(container):
+        raise ValueError('You can not add the copied content here.')